Move GridSiteSyncSubmitH API data source from live DB to local DB
- we take a regular pull from the live DB and place it in the local DB
  via cron JOB.
